+++ dom.js
@@ -1,8 +1,8 @@
-var DOM = require('./src/cssselector');
+var DOM = require('./src/main/javascript/cssselector');
 
 var Core = require('coreutil/core');
-var RS = require('./src/domresultset');
-var Attr = require('./src/cssattribute');
+var RS = require('./src/main/javascript/domresultset');
+var Attr = require('./src/main/javascript/cssattribute');
 
 Core.extend(Core, RS);
 Core.extend(Core, Attr);
