+++ src/main/javascript/cssoperators.js
