+++ src/test/javascript/test.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>DOM Test</title>
 
-    <script src="../dist/dom.bundle.js"></script>
+    <script src="../../../dist/dom.bundle.js"></script>
 </head>
 <body>
